pyproject.toml
[project]
name = "golf-swing-ai-coach"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "nvidia-cublas-cu12>=12.9.1.4",
    "nvidia-cuda-nvrtc-cu12>=12.9.86",
    "nvidia-cuda-runtime-cu12>=12.9.79",
    "nvidia-cudnn-cu12>=9.19.0.56",
    "nvidia-cufft-cu12>=11.4.1.4",
    "nvidia-curand-cu12>=10.3.10.19",
    "nvidia-nvjitlink-cu12>=12.9.86",
    "onnxruntime-gpu>=1.24.1",
    "opencv-python>=4.8.0",
    "pillow>=12.1.1",
    "rtmlib>=0.0.15",
]
